+++ smtlib-backends-z3/CHANGELOG.md
@@ -9,6 +9,7 @@ time
   - what happens when sending an empty command
   - what happens when sending a command not producing any output
 - **(breaking change)** removed `Data.Default` instance of `Config`
+- dropped dependency on `inline-c`
 
 # v0.2
 - make test-suite compatible with `smtlib-backends-0.2`